nit: move files around
