--- data/lib/src/mock/mock_data.dart
@@ -1,1 +1,0 @@
-
